ft(homepage): implement homepage
