+++ src/components/Characters/index.scss
@@ -2,16 +2,6 @@
   width: 90%;
   margin: 0 auto;
 
-  .character-title {
-    padding: 30px 0;
-
-    hr.title-line {
-      height: 3px;
-      background: #000;
-      max-width: 10%;
-    }
-  }
-
   .characters {
     display: grid;
     grid-template-columns: repeat(2, minmax(0, 1fr));
