+++ src/components/Home/index.scss
@@ -1,0 +1,19 @@
+.home {
+  .home-content-wrap {
+    width: 85%;
+    margin: 0 auto;
+    .starship {
+      display: grid;
+      grid-template-columns: repeat(3, minmax(0, 1fr));
+      grid-gap: 40px 0;
+      justify-items: center;
+    }
+
+    .characters {
+      display: grid;
+      grid-template-columns: repeat(2, minmax(0, 1fr));
+      grid-gap: 40px 0;
+      justify-items: center;
+    }
+  }
+}
