+++ src/components/shared/Cards/HorizontalCard/index.jsx
@@ -3,8 +3,8 @@ import Card from 'react-bootstrap/Card';
 import char1 from '../../../../images/character-1.jpg';
 import './index.scss';
 
-const HorizontalCard = ({ title, birthYear, gender }) => (
-  <Card style={{ width: '35rem' }} className="horizontal-card">
+const HorizontalCard = ({ title, birthYear, gender, width }) => (
+  <Card style={{ width }} className="horizontal-card">
     <Card.Img variant="top" src={char1} />
     <Card.Body>
       <Card.Title>{title}</Card.Title>
