+++ src/components/shared/StarWar/StarwarTitle.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import './index.scss';
+
+const StarwarTitle = ({ header }) => {
+  return (
+    <div className="starship-title">
+      <h3>{header}</h3>
+      <hr className="title-line" />
+    </div>
+  );
+};
+
+export default StarwarTitle;
