+++ src/components/shared/StarWar/index.scss
@@ -1,0 +1,9 @@
+.starship-title {
+  padding: 30px 0;
+
+  hr.title-line {
+    height: 3px;
+    background: #000;
+    max-width: 5%;
+  }
+}
